support for inner queries alias without the `AS`.
Right now a query like this will be marked as unknow:
      SELECT foo.abc FROM (select abc from bar) foo;
This is a valid syntax, as the AS is not mandatory. I tested this on
MySQL and SQLite3, and it is valid syntax on both.

Seems like the parser lemon file was missing the codition for (...) ID ,
so I added it, and that query now is parsed correctly.